+++ packages/gator-permissions-snap/src/core/types.ts
@@ -165,7 +165,6 @@ export type LifecycleOrchestrationHandlers<
     updateContext: (updateContextArgs: {
       updatedContext: TContext;
     }) => Promise<void>;
-    isAdjustmentAllowed: boolean;
   }) => void;
 
   /**
@@ -192,9 +191,7 @@ export type RuleData = {
   iconData?: IconData | undefined;
   error?: string | undefined;
   options?: string[] | undefined;
-  isAdjustmentAllowed: boolean;
-  /** For rules types that are defined in the 7715 specification, this will be true. And, are always allowed to be adjusted. */
-  is7715RuleType?: boolean;
+  isEditable: boolean;
   /** For datetime rules: whether to disable selection of past dates. Defaults to false. */
   allowPastDate?: boolean | undefined;
 };